v0.6.3.4
-更新制作语言包中

+++ UI/Settings/GameSetHelper.cs
@@ -44,12 +44,12 @@ namespace KFrame.UI
         public static void SetLanguage(LanguageType languageType)
         {
             CurSettingsData.Language = languageType;
-            LocalizationSystem.LanguageType = languageType;
+            LocalizationSystem.LanguageType = (int)languageType;
         }
 
         public static void LoadGameSettings()
         {
-            LocalizationSystem.LanguageType = CurSettingsData.Language;
+            LocalizationSystem.LanguageType = (int)CurSettingsData.Language;
         }
     }
 }
